Implement node stats
Nodes         dev2n1        dev2n2        dev2n3
 score      | 94            95            95
  load15m   | 0.1           0.0           0.1
  mem       | 26/100%:1.93g 28/100%:1.93g 28/100%:1.93g
  swap      | 6/100%:1.9g   1/100%:1.9g   1/100%:1.9g

+++ core/cluster/cluster_node.go
@@ -34,7 +34,7 @@ type (
 		Load15M      float64 `json:"load_15m"`
 		MemAvailPct  uint64  `json:"mem_avail"`
 		MemTotalMB   uint64  `json:"mem_total"`
-		Score        uint    `json:"score"`
+		Score        uint64  `json:"score"`
 		SwapAvailPct uint64  `json:"swap_avail"`
 		SwapTotalMB  uint64  `json:"swap_total"`
 	}
